whops forgot to make facebook uids long

+++ letterstocrushes.Core/Interfaces/IQueryLetters.cs
@@ -11,7 +11,7 @@ namespace letterstocrushes.Core.Interfaces
     {
         void hideLetter(int lucky_id, string userip, string cookie_value, string user_name, bool is_user_mod);
         void unhideLetter(int lucky_id, string userip, string cookie_value, string user_name, bool is_user_mod);
-        void facebookLetter(int lucky_id, int toFacebookUID, int fromFacebookUID, string userip, string cookie_value, string user_name, bool is_user_mod);
+        void facebookLetter(int lucky_id, long toFacebookUID, long fromFacebookUID, string userip, string cookie_value, string user_name, bool is_user_mod);
         Boolean editLetter(int letter_id, string new_letter, string userip, string cookie_value, string user_name, bool is_user_mod);
         Letter getLetter(int id);
         int getLetterCount();
